Rename files according to new url
