+++ config/webpack.config.js
@@ -27,7 +27,7 @@ const entryFiles = {
   'no-react': './src/js/no-react-entry.js',
   'user-profile': './src/js/user-profile/user-profile-entry.jsx',
   auth: './src/js/auth/auth-entry.jsx',
-  'va-letters': './src/js/va-letters/va-letters-entry.jsx',
+  letters: './src/js/letters/letters-entry.jsx',
   pensions: './src/js/pensions/pensions-entry.jsx',
   'post-911-gib-status': './src/js/post-911-gib-status/post-911-gib-status-entry.jsx',
   'pre-need': './src/js/pre-need/pre-need-entry.jsx',
