+++ src/js/letters/utils/helpers.js
