+++ test/letters/containers/LettersApp.unit.spec.jsx
@@ -2,17 +2,17 @@ import React from 'react';
 import SkinDeep from 'skin-deep';
 import { expect } from 'chai';
 
-import VALettersApp from '../../../src/js/va-letters/containers/VALettersApp';
+import LettersApp from '../../../src/js/letters/containers/LettersApp';
 
-import reducer from '../../../src/js/va-letters/reducers';
+import reducer from '../../../src/js/letters/reducers';
 import createCommonStore from '../../../src/js/common/store';
 
 const store = createCommonStore(reducer);
 
-describe('<VALettersApp>', () => {
+describe('<LettersApp>', () => {
   it('should render', () => {
     const mockRoutes = [{ path: '/fake' }];
-    const tree = SkinDeep.shallowRender(<VALettersApp store={store} location={{ pathname: '/blah' }} route={{ childRoutes: mockRoutes }}/>);
+    const tree = SkinDeep.shallowRender(<LettersApp store={store} location={{ pathname: '/blah' }} route={{ childRoutes: mockRoutes }}/>);
     const vdom = tree.getRenderOutput();
     expect(vdom).to.not.be.undefined;
   });
